테이블 수정 : Measurement 필드 타입 Double로 설정

+++ src/main/java/fittering/mall/controller/dto/request/RequestMeasurementDto.java
@@ -8,14 +8,14 @@ import lombok.*;
 @AllArgsConstructor
 public class RequestMeasurementDto {
     @NonNull
-    private Integer height;
+    private Double height;
     @NonNull
-    private Integer weight;
-    private Integer arm;
-    private Integer leg;
-    private Integer shoulder;
-    private Integer waist;
-    private Integer chest;
-    private Integer thigh;
-    private Integer hip;
+    private Double weight;
+    private Double arm;
+    private Double leg;
+    private Double shoulder;
+    private Double waist;
+    private Double chest;
+    private Double thigh;
+    private Double hip;
 }
